prettying up text; directly casting ints as floats

+++ src/main/java/com/lvl6/server/controller/AttackClanRaidMonsterController.java
@@ -86,7 +86,9 @@ import com.lvl6.utils.utilmethods.UpdateUtils;
 
   @Override
   protected void processRequestEvent(RequestEvent event) throws Exception {
-    AttackClanRaidMonsterRequestProto reqProto = ((AttackClanRaidMonsterRequestEvent)event).getAttackClanRaidMonsterRequestProto();
+    AttackClanRaidMonsterRequestProto reqProto =
+    		((AttackClanRaidMonsterRequestEvent)event)
+    		.getAttackClanRaidMonsterRequestProto();
     log.info("reqProto=");
     log.info(reqProto +"");
 
@@ -101,14 +103,20 @@ import com.lvl6.utils.utilmethods.UpdateUtils;
     int damageDealt = reqProto.getDamageDealt(); //remember take min of this with monsters's remaining hp
     
     //extract the new healths for the monster(s)
-    List<UserMonsterCurrentHealthProto> monsterHealthProtos = reqProto.getMonsterHealthsList();
-    Map<Long, Integer> userMonsterIdToExpectedHealth = new HashMap<Long, Integer>();
-    MonsterStuffUtils.getUserMonsterIds(monsterHealthProtos, userMonsterIdToExpectedHealth);
+    List<UserMonsterCurrentHealthProto> monsterHealthProtos =
+    		reqProto.getMonsterHealthsList();
+    Map<Long, Integer> userMonsterIdToExpectedHealth =
+    		new HashMap<Long, Integer>();
+    MonsterStuffUtils.getUserMonsterIds(monsterHealthProtos,
+    		userMonsterIdToExpectedHealth);
     
-    FullUserMonsterProto userMonsterThatAttacked = reqProto.getUserMonsterThatAttacked();
-    UserCurrentMonsterTeamProto userMonsterTeam = reqProto.getUserMonsterTeam();
+    FullUserMonsterProto userMonsterThatAttacked =
+    		reqProto.getUserMonsterThatAttacked();
+    UserCurrentMonsterTeamProto userMonsterTeam =
+    		reqProto.getUserMonsterTeam();
     
-    AttackClanRaidMonsterResponseProto.Builder resBuilder = AttackClanRaidMonsterResponseProto.newBuilder();
+    AttackClanRaidMonsterResponseProto.Builder resBuilder =
+    		AttackClanRaidMonsterResponseProto.newBuilder();
     resBuilder.setStatus(AttackClanRaidMonsterStatus.FAIL_OTHER);
     resBuilder.setSender(sender);
     resBuilder.setUserMonsterThatAttacked(userMonsterThatAttacked);
@@ -740,17 +748,17 @@ import com.lvl6.utils.utilmethods.UpdateUtils;
   	
   	int userId = cepfu.getUserId();
   	int userCrsDmg = cepfu.getCrsDmgDone() + cepfu.getCrsmDmgDone();
-  	float userCrsContribution = (float) ((float) userCrsDmg / (float) stageHp);
+  	float userCrsContribution = (float) ((float) userCrsDmg) / ((float) stageHp);
   	
   	int staticDataId = 0;
   	//TODO: FIGURE OUT IF CURRENCY CALCULATION IS OK (right now just truncating the value) 
   	//create the userOilReward maybe
-  	int userOilReward = (int) (reward.getOilDrop() * userCrsContribution);
+  	int userOilReward = (int) (((float)reward.getOilDrop()) * userCrsContribution);
   	createClanEventPersistentUserReward(MiscMethods.OIL, userOilReward, staticDataId,
   			crsId, crsStartDate, crsEndDate, clanEventId, userId, userRewards);
   	
   	//create the userOilReward maybe  	
-  	int userCashReward = (int) (reward.getCashDrop() * userCrsContribution);
+  	int userCashReward = (int) (((float)reward.getCashDrop()) * userCrsContribution);
   	createClanEventPersistentUserReward(MiscMethods.CASH, userCashReward, staticDataId,
   			crsId, crsStartDate, crsEndDate, clanEventId, userId, userRewards);
 
@@ -761,7 +769,8 @@ import com.lvl6.utils.utilmethods.UpdateUtils;
   	}
   	
   	//compute monster reward
-  	float monsterDropRate = userCrsContribution * reward.getExpectedMonsterRewardQuantity();
+  	float monsterDropRate = userCrsContribution * 
+  			((float)reward.getExpectedMonsterRewardQuantity());
   	Random rand = reward.getRand();
   	if (rand.nextFloat() < monsterDropRate) {
   		//user gets the monster reward
